Add files via upload
TASK 1:-  BASIC CHATBOT 
It's my first task completed under the tasks given by CodeAlpha.
It's a basic chat bot created using python.
Library included : NLTK
Description :
it is a basic chat bot in which a pair of statements a feeded  i.e. when a second person calls out one statement of the pair the bot replies the second one.
Additional pairs and tasks can be added in it for various needs.